+++ docs/search/files_9.js
@@ -7,6 +7,7 @@ var searchData=
   ['moduleframeworkserviceprovider_2ephp',['ModuleFrameworkServiceProvider.php',['../ModuleFrameworkServiceProvider_8php.html',1,'']]],
   ['moduleinactive_2ephp',['ModuleInactive.php',['../ModuleInactive_8php.html',1,'']]],
   ['moduleinstance_2ephp',['ModuleInstance.php',['../Contracts_2ModuleInstance_8php.html',1,'(Global Namespace)'],['../ModuleInstance_8php.html',1,'(Global Namespace)']]],
+  ['moduleinstancedisabled_2ephp',['ModuleInstanceDisabled.php',['../ModuleInstanceDisabled_8php.html',1,'']]],
   ['moduleinstanceevaluator_2ephp',['ModuleInstanceEvaluator.php',['../Contracts_2Evaluator_2ModuleInstanceEvaluator_8php.html',1,'(Global Namespace)'],['../Evaluator_2ModuleInstanceEvaluator_8php.html',1,'(Global Namespace)'],['../Facade_2ModuleInstanceEvaluator_8php.html',1,'(Global Namespace)']]],
   ['moduleinstancegroupoverridepermission_2ephp',['ModuleInstanceGroupOverridePermission.php',['../ModuleInstanceGroupOverridePermission_8php.html',1,'']]],
   ['moduleinstancepermission_2ephp',['ModuleInstancePermission.php',['../ModuleInstancePermission_8php.html',1,'']]],