+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.1.1] - (18/03/2020)
+
 ### Added
 - Middleware to throw an ActivityDisabled exception if the activity is disabled
 - Middleware to throw a ModuleInstanceDisabled exception if the module instance is disabled
@@ -120,7 +122,8 @@ BristolSU\Support\Testing\HandlesAuthorization to help with testing modules with
 ### Added
 - Initial Release
 
-[Unreleased]: https://github.com/bristol-su/support/compare/v3.1...HEAD
+[Unreleased]: https://github.com/bristol-su/support/compare/v3.1.1...HEAD
+[3.1.1]: https://github.com/bristol-su/support/compare/v3.1...v3.1.1
 [3.1]: https://github.com/bristol-su/support/compare/v3.0.2...v3.1
 [3.0.2]: https://github.com/bristol-su/support/compare/v3.0.1...v3.0.2
 [3.0.1]: https://github.com/bristol-su/support/compare/v3.0...v3.0.1